+++ src/decorator/options/RelationOptions.ts
@@ -1,3 +1,4 @@
+import {DeferrableType} from "../../metadata/types/DeferrableType";
 import {OnDeleteType} from "../../metadata/types/OnDeleteType";
 import {OnUpdateType} from "../../metadata/types/OnUpdateType";
 
@@ -29,6 +30,11 @@ export interface RelationOptions {
      * Database cascade action on update.
      */
     onUpdate?: OnUpdateType;
+
+    /**
+     * Indicate if foreign key constraints can be deferred.
+     */
+    deferrable?: DeferrableType;
 
     /**
      * Indicates if this relation will be a primary key.