+++ src/metadata-builder/RelationJoinColumnBuilder.ts
@@ -70,6 +70,7 @@ export class RelationJoinColumnBuilder {
             referencedColumns: referencedColumns,
             onDelete: relation.onDelete,
             onUpdate: relation.onUpdate,
+            deferrable: relation.deferrable,
         });
 
         // Oracle does not allow both primary and unique constraints on the same column