+++ src/metadata/types/DeferrableType.ts
@@ -1,0 +1,5 @@
+
+/**
+ * DEFERRABLE type to be used to specify if foreign key constraints can be deferred.
+ */
+export type DeferrableType = "INITIALLY IMMEDIATE"|"INITIALLY DEFERRED";