+++ src/query-builder/transformer/DocumentToEntityTransformer.ts
@@ -98,7 +98,7 @@ export class DocumentToEntityTransformer {
                 } else {
                     embedded.columns.forEach(column => {
                         const value = document[embedded.prefix][column.databaseNameWithoutPrefixes];
-                        if (!value) return;
+                        if (value === undefined) return;
 
                         if (!entity[embedded.propertyName])
                             entity[embedded.propertyName] = embedded.create();